adding B Problem in 11 segment basic-cpp tlx

+++ TLX/Course Pemrograman Dasar C++/11. Analisis Kompleksitas/B_Perkenalan_Format_Soal_Pemrograman_Kompetitif_I.cpp
@@ -3,13 +3,13 @@
 using namespace std;
 int main(){
     int N;cin >> N;
-    vector <int> A;
+    vector <long long> A;
     for (int i = 0; i < N; i++) {
         int x;
         cin >> x;
         A.push_back(x);
     }
-    int jumlah = 0;
+    long long jumlah = 0;
     for (int i = 0; i < N; i++) {
         jumlah += A[i];
     }
